test: add integration test project
Adds an xUnit project for writing integration tests. This was chosen at this stage as it's quicker and easier than trying to write unit tests due to that nature of needing a storage to interface with.

A test has been added that would have caught the issue in the previous release where the CascFindData was being populated incorrectly.