docs: add docfx [skip ci]
Sets up DocFX for automatic documentation generation, and adds relevant pages that the site requires besides the automatically generated ones.

Minor corrections made to the README to ensure the documentation is correct.

Adds GitHub workflow to automatically publish the DocFX site to GitHub pages for the CascLib.NET repository when the main branch is pushed.